Attempt to use gui in main file
Unfortunately, the GUI file's constructor code is locked in the appDesigner. I'm sure I could somehow pull the code into a different text editor, edit there and use that, but it seems like too much of a hassle, especially to fix if something needs to be added.

My solution is just to use the GUI file as it is in order to save time. It's not good coding practice, but it shouldn't matter for the scope of this project.